Changed app. init. and final. to be like unittest
The established tradition of the initialization and finalization methods
being called ``setUp`` and ``tearDown`` is now reflected and perpetuated
for the sake of sanity and consistency with already existing nomenclature
for such methods within the context of ``fixtures``. In this case, the
application framework is providing the scaffolding for the user
implemented application fixture.